limpiando

+++ lab8/ejercicio/src/ejercicio3/test_interface.java
@@ -2,5 +2,5 @@ package ejercicio3;
 
 interface test_interface {
     void calcular();
-    Integer logica(Integer ... list);
+    Integer logica(Float ... list);
 }
